requirements.txt
Django==1.8.1
GnuPGInterface==0.3.2
MySQL-python==1.2.3
PIL==1.1.7
Pillow==2.8.1
Restler==0.7.1
SQLAlchemy==1.0.4
apt-xapian-index==0.44
argparse==1.2.1
boto==2.38.0
captricity-python-client==0.20
cffi==1.1.0
chardet==2.0.1
command-not-found==0.2.44
configobj==4.7.2
cryptography==0.9
decorator==3.3.2
## FIXME: could not find svn URL in dependency_links for this package:
distribute==0.6.24dev-r0
django-bootstrap3==5.4.0
django-cors-headers==1.1.0
django-extensions==1.5.5
django-s3direct==0.3.11
django-storages==1.1.8
enum34==1.0.4
idna==2.0
imutils==0.2.1
ipaddress==1.0.7
ipython==0.12.1
language-selector==0.1
logger==1.4
matplotlib==1.1.1rc
ndg-httpsclient==0.4.0
nose==1.1.2
numexpr==1.4.2
numpy==1.9.2
pandas==0.7.0
pexpect==2.3
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.13
pycurl==7.19.0
pygraphviz==1.3rc2
pyjade==3.0.0
pyparsing==1.5.2
python-apt==0.8.3ubuntu7.3
python-dateutil==1.5
python-debian==0.1.21ubuntu1
python-magic==0.4.6
pytz==2011k
pyzmq==2.1.11
requests==2.7.0
rpy==1.0.3
scikits.statsmodels==0.3.1
scipy==0.9.0
simplegeneric==0.7
six==1.9.0
sympy==0.7.1.rc1
tables==2.3.1
tornado==2.1
ufw==0.31.1-1
wsgiref==0.1.2
xlrd==0.6.1
yapf==0.1.9
